requirements.txt
Faker==30.8.2
Flask==3.1.0
python-dotenv==1.0.1
SQLAlchemy==2.0.36
